Booking view
Now the booking page could be accessed through a book now button when we are logged in. I still new to do the redirection when the user is not logged. Need to do some changes because there is an exception when I try to click the button for searching available rooms